# tixmoapi2/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TixMo API",
    "version": "0.1.0",
    "description": "Enterprise-grade ticket vendor platform API for event promoters",
    "contact": {
      "name": "TixMo Team",
      "email": "api@tixmo.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.tixmo.com/v1",
      "description": "Production server (future)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token"
      }
    },
...
  },
...
  "paths": {
...
    "/organizations": {
      "post": {
        "summary": "Create a new organization",
        "description": "Create a new organization (promoter, venue, or reseller). Requires ADMIN or PROMOTER role.",
        "tags": [
          "Organizations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "slug",
                  "type"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Awesome Events Co"
                  },
                  "slug": {
                    "type": "string",
                    "example": "awesome-events",
                    "description": "URL-friendly identifier (lowercase, numbers, hyphens only)"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "PROMOTER",
                      "VENUE",
                      "RESELLER"
                    ],
                    "example": "PROMOTER"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organization created successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Requires ADMIN or PROMOTER role"
          },
          "409": {
            "description": "Organization with this slug already exists"
          }
        }
      },
      "get": {
        "summary": "List all organizations",
        "description": "Get a paginated list of organizations with optional filters",
        "tags": [
          "Organizations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "PROMOTER",
                "VENUE",
                "RESELLER"
              ]
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "SUSPENDED",
                "PENDING"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organizations retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/organizations/{id}": {
      "get": {
        "summary": "Get organization by ID",
        "description": "Get detailed information about an organization including members and stats",
        "tags": [
          "Organizations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Organization retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Organization not found"
          }
        }
      },
      "put": {
        "summary": "Update organization",
        "description": "Update organization details. Must be admin or organization member.",
        "tags": [
          "Organizations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Updated Events Co"
                  },
                  "slug": {
                    "type": "string",
                    "example": "updated-events"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "ACTIVE",
                      "SUSPENDED",
                      "PENDING"
                    ],
                    "example": "ACTIVE"
                  },
                  "stripeAccountId": {
                    "type": "string",
                    "example": "acct_123456789"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Organization updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Organization not found"
          },
          "409": {
            "description": "Slug already exists"
          }
        }
      },
      "delete": {
        "summary": "Delete organization",
        "description": "Delete an organization (admin only). Cannot delete if organization has events or venues.",
        "tags": [
          "Organizations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Organization deleted successfully"
          },
          "400": {
            "description": "Cannot delete organization with existing events/venues"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Admin only"
          },
          "404": {
            "description": "Organization not found"
          }
        }
      }
    },
    "/organizations/{id}/members": {
      "post": {
        "summary": "Add member to organization",
        "description": "Add a user as a member of the organization",
        "tags": [
          "Organizations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Member added successfully"
          },
          "400": {
            "description": "User is already a member"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Organization or user not found"
          }
        }
      }
    },
    "/organizations/{id}/members/{userId}": {
      "delete": {
        "summary": "Remove member from organization",
        "description": "Remove a user from the organization",
        "tags": [
          "Organizations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member removed successfully"
          },
          "400": {
            "description": "User is not a member of this organization"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Organization or user not found"
          }
        }
      }
    },
...
  }
}